test autoritydiscriminationgame, recordgame and gameHelper.ts
